another attempt.

+++ src/common.h
@@ -7,25 +7,25 @@ static const int MAX_STRING_LENGTH = 1000;
 static const int SEED =	1159241;
 static const int TSIZE = 1048576;
 
-typedef long double real;
+typedef double real;
 
 typedef struct cooccur_rec {
-    long int word1;
-    long int word2;
-    real val;
+  long long word1;
+  long long word2;
+  real val;
 } CREC;
 
 typedef struct cooccur_rec_id {
-    long int word1;
-    long int word2;
-    real val;
-    long int id;
+  long long word1;
+  long long word2;
+  real val;
+  long long id;
 } CRECID;
 
 typedef struct hashrec {
-    char *word;
-    long long id;
-    struct hashrec *next;
+  char *word;
+  long long id;
+  struct hashrec *next;
 } HASHREC;
 
 /* Efficient string comparison */
